Move hooks outside of shared folder
